+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/HuffElement.kt
@@ -1,0 +1,8 @@
+package com.github.com.cakevm.intellij_huff_plugin.language
+
+import com.intellij.psi.PsiElement
+import com.intellij.psi.PsiReference
+
+interface HuffElement : PsiElement {
+  override fun getReference(): PsiReference?
+}