+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/HuffLanguage.kt
@@ -5,6 +5,8 @@ import com.intellij.lang.Language
 class HuffLanguage : Language(NAME) {
   override fun getDisplayName() = NAME
 
+  override fun isCaseSensitive() = true
+
   companion object {
     const val NAME = "Huff"
     @JvmStatic val INSTANCE = HuffLanguage()