+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/psi/element/HuffMacroDefinitionElement.kt
@@ -1,0 +1,6 @@
+package com.github.com.cakevm.intellij_huff_plugin.language.psi.element
+
+import com.github.com.cakevm.intellij_huff_plugin.language.HuffNamedElement
+import com.github.com.cakevm.intellij_huff_plugin.language.types.HuffMember
+
+interface HuffMacroDefinitionElement : HuffNamedElement, HuffMember