+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/psi/stub/HuffNamedStub.kt
@@ -1,0 +1,5 @@
+package com.github.com.cakevm.intellij_huff_plugin.language.psi.stub
+
+interface HuffNamedStub {
+  val name: String?
+}