+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/reference/HuffReference.kt
@@ -1,0 +1,13 @@
+package com.github.com.cakevm.intellij_huff_plugin.language.reference
+
+import com.github.com.cakevm.intellij_huff_plugin.language.HuffElement
+import com.intellij.psi.PsiElement
+import com.intellij.psi.PsiPolyVariantReference
+
+interface HuffReference : PsiPolyVariantReference {
+  override fun getElement(): HuffElement
+
+  override fun resolve(): HuffElement?
+
+  fun multiResolve(): Collection<PsiElement>
+}