+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/types/HuffMember.kt
@@ -1,0 +1,5 @@
+package com.github.com.cakevm.intellij_huff_plugin.language.types
+
+interface HuffMember {
+  fun getName(): String?
+}